refactor ica interaction
