+++ src/dragen_align_pa/ica_utils.py
@@ -1,224 +1,29 @@
 """
-This module centralizes all interactions with the Illumina Connected Analytics (ICA)
-API and CLI. It provides helper functions for authentication, file/folder operations,
-pipeline status checking, and data streaming.
+This module provides high-level helper functions and business logic for
+interacting with ICA. It orchestrates calls to the low-level ica_api
+and ica_cli modules.
 """
 
-import contextlib
 import hashlib
 import json
-import os
-from collections.abc import Iterator
-from typing import TYPE_CHECKING, Any, Final, Literal
-
-if TYPE_CHECKING:
-    from collections.abc import Sequence
-    from subprocess import CompletedProcess
+from typing import TYPE_CHECKING
 
 import cpg_utils
 import icasdk
 import requests
 from google.cloud import exceptions as gcs_exceptions
-from google.cloud import secretmanager
-from google.cloud.storage.bucket import Bucket
-from icasdk import ApiClient, Configuration
-from icasdk.apis.tags import project_analysis_api, project_data_api
-from icasdk.exceptions import ApiException
 from icasdk.model.create_data import CreateData
 from loguru import logger
 
-from dragen_align_pa import utils
-from dragen_align_pa.constants import ICA_CLI_SETUP, ICA_REST_ENDPOINT
+from dragen_align_pa import ica_api_utils
 
-# --- Secret Management ---
-
-SECRET_CLIENT: Final = secretmanager.SecretManagerServiceClient()
-SECRET_PROJECT: Final = 'cpg-common'
-SECRET_NAME: Final = 'illumina_cpg_workbench_api'
-SECRET_VERSION: Final = 'latest'
-
-
-def get_ica_secrets() -> dict[Literal['projectID', 'apiKey'], str]:
-    """Gets the project ID and API key used to interact with ICA
-
-    Returns:
-        dict[str, str]: A dictionary with the keys projectId and apiKey
-    """
-    secret_path: str = SECRET_CLIENT.secret_version_path(
-        project=SECRET_PROJECT,
-        secret=SECRET_NAME,
-        secret_version=SECRET_VERSION,
-    )
-    response: secretmanager.AccessSecretVersionResponse = SECRET_CLIENT.access_secret_version(
-        request={'name': secret_path},
-    )
-    return json.loads(response.payload.data.decode('UTF-8'))
-
-
-@contextlib.contextmanager
-def get_ica_api_client() -> Iterator[ApiClient]:
-    """
-    Provides a context-managed icasdk.ApiClient.
-    Handles fetching secrets, configuring, and closing the client.
-    """
-    secrets: dict[Literal['projectID', 'apiKey'], str] = get_ica_secrets()
-    api_key: str = secrets['apiKey']
-
-    configuration = Configuration(host=ICA_REST_ENDPOINT)
-    configuration.api_key['ApiKeyAuth'] = api_key
-
-    with ApiClient(configuration=configuration) as api_client:
-        try:
-            yield api_client
-        except ApiException as e:
-            logger.error(f'ICA API Exception caught by context manager: {e}')
-            raise
-        except Exception as e:
-            logger.error(f'Non-API Exception caught by context manager: {e}')
-            raise
-
-
-def find_ica_file_path_by_name(parent_folder: str, file_name: str) -> str:
-    """
-    Finds a file in ICA using the CLI and returns its full `details.path`.
-    """
-    command = [
-        'icav2',
-        'projectdata',
-        'list',
-        '--parent-folder',
-        parent_folder,
-        '--data-type',
-        'FILE',
-        '--file-name',
-        file_name,
-        '--match-mode',
-        'EXACT',
-        '-o',
-        'json',
-    ]
-    result: CompletedProcess[Any] = utils.run_subprocess_with_log(command, f'Find ICA file {file_name}')
-    try:
-        data = json.loads(result.stdout)
-        if not data.get('items'):
-            raise ValueError(
-                f'No file found with name "{file_name}" in folder "{parent_folder}"',
-            )
-
-        file_path = data['items'][0].get('details', {}).get('path')
-        if not file_path:
-            raise ValueError(
-                f'File "{file_name}" found, but it has no "details.path" in API response.',
-            )
-
-        logger.info(f'Found {file_name} at path: {file_path}')
-        return file_path
-
-    except json.JSONDecodeError:
-        logger.error(f'Failed to decode JSON from icav2 list command: {result.stdout}')
-        raise
-    except (ValueError, IndexError) as e:
-        logger.error(f'Error parsing icav2 list output for {file_name}: {e}')
-        raise
-
-
-# --- Pipeline Management (SDK) ---
-
-
-def check_ica_pipeline_status(
-    api_instance: project_analysis_api.ProjectAnalysisApi,
-    path_params: dict[str, str],
-) -> str:
-    """Check the status of an ICA pipeline via a pipeline ID
-
-    Args:
-        api_instance (project_analysis_api.ProjectAnalysisApi): An instance of the ProjectAnalysisApi
-        path_params (dict[str, str]): Dict with projectId and analysisId
-
-    Raises:
-        icasdk.ApiException: Any exception if the API call is incorrect
-
-    Returns:
-        str: The status of the pipeline. Can be one of ['REQUESTED', 'AWAITINGINPUT', 'INPROGRESS', 'SUCCEEDED', 'FAILED', 'FAILEDFINAL', 'ABORTED']
-    """  # noqa: E501
-    try:
-        api_response = api_instance.get_analysis(path_params=path_params)  # type: ignore[ReportUnknownVariableType]
-        pipeline_status: str = api_response.body['status']  # type: ignore[ReportUnknownVariableType]
-        return pipeline_status  # type: ignore[ReportUnknownVariableType]
-    except icasdk.ApiException as e:
-        raise icasdk.ApiException(
-            f'Exception when calling ProjectAnalysisApi -> get_analysis: {e}',
-        ) from e
-
-
-# --- Data Operations (SDK) ---
-
-
-def check_object_already_exists(
-    api_instance: project_data_api.ProjectDataApi,
-    path_params: dict[str, str],
-    file_name: str,
-    folder_path: str,
-    object_type: str,
-) -> tuple[str, str] | None:
-    """Check if an object already exists in ICA.
-
-    Args:
-        api_instance (project_data_api.ProjectDataApi): An instance of the ProjectDataApi
-        path_params (dict[str, str]): A dict with the projectId
-        file_name (str): The name of the object that you want to check in ICA e.g.
-        folder_path (str): The path to the object that you want to create in ICA.
-        object_type (str): The type of hte object to create in ICA. Must be one of ['FILE', 'FOLDER']
-
-    Raises:
-        NotImplementedError: Only checks for files with the status 'PARTIAL' or 'AVAILABLE'
-        icasdk.ApiException: Other API errors
-
-    Returns:
-        tuple[str, str] | None: (object_ID, object_status) if it exists, or else None
-    """
-    query_params: dict[str, Sequence[str] | list[str] | str] = {
-        'filePath': [f'{folder_path}/{file_name}'],
-        'filePathMatchMode': 'STARTS_WITH_CASE_INSENSITIVE',
-        'type': object_type,
-    }
-    if object_type == 'FILE':
-        query_params = {  # pyright: ignore[reportUnknownVariableType]
-            'filename': [file_name],
-            'filenameMatchMode': 'EXACT',
-        } | query_params
-    logger.info(
-        f'Checking to see if the {object_type} object already exists at {folder_path}/{file_name}',
-    )
-    try:
-        api_response = api_instance.get_project_data_list(  # type: ignore[ReportUnknownVariableType]
-            path_params=path_params,  # type: ignore[ReportUnknownVariableType]
-            query_params=query_params,  # type: ignore[ReportUnknownVariableType]
-        )  # type: ignore[ReportUnknownVariableType]
-
-        if len(api_response.body['items']) == 0:  # type: ignore[ReportUnknownVariableType]
-            return None
-
-        object_data = api_response.body['items'][0]['data']  # pyright: ignore[reportUnknownVariableType]
-        object_id = object_data['id']  # pyright: ignore[reportUnknownVariableType]
-        status: str = object_data['details'].get('status', 'UNKNOWN')  # pyright: ignore[reportUnknownVariableType]
-
-        if object_type == 'FOLDER':
-            return object_id, status  # Folders have status, e.g., 'AVAILABLE'
-
-        if status in ('PARTIAL', 'AVAILABLE'):
-            return object_id, status
-
-        # Statuses are ["PARTIAL", "AVAILABLE", "ARCHIVING", "ARCHIVED", "UNARCHIVING", "DELETING", ]
-        raise NotImplementedError(f'Checking for file status "{status}" is not implemented yet.')
-    except icasdk.ApiException as e:
-        raise icasdk.ApiException(
-            f'Exception when calling ProjectDataApi -> get_project_data_list: {e}',
-        ) from e
+if TYPE_CHECKING:
+    from google.cloud.storage.bucket import Bucket
+    from icasdk.apis.tags import project_data_api
 
 
 def create_upload_object_id(
-    api_instance: project_data_api.ProjectDataApi,
+    api_instance: 'project_data_api.ProjectDataApi',
     path_params: dict[str, str],
     sg_name: str,
     file_name: str,
@@ -243,7 +48,7 @@ def create_upload_object_id(
         tuple[str, str]: (object_ID, status)
         Status will be from ICA, e.g. 'AVAILABLE', 'PARTIAL'.
     """
-    existing_object_details: tuple[str, str] | None = check_object_already_exists(
+    existing_object_details: tuple[str, str] | None = ica_api_utils.check_object_already_exists(
         api_instance=api_instance,
         path_params=path_params,
         file_name=file_name,
@@ -284,55 +89,8 @@ def create_upload_object_id(
         ) from e
 
 
-def find_file_id_by_name(
-    api_instance: project_data_api.ProjectDataApi,
-    path_parameters: dict[str, str],
-    parent_folder_path: str,
-    file_name: str,
-) -> str:
-    """
-    Finds a specific file ID in an ICA folder by its exact name.
-    (Used by download_specific_files_from_ica.py)
-    """
-    logger.info(f"Searching for file '{file_name}' in '{parent_folder_path}'...")
-    try:
-        api_response = api_instance.get_project_data_list(  # pyright: ignore[reportUnknownVariableType]
-            path_params=path_parameters,
-            query_params={  # pyright: ignore[reportUnknownVariableType]
-                'parentFolderPath': parent_folder_path,
-                'filename': [file_name],
-                'filenameMatchMode': 'EXACT',
-                'pageSize': '2',
-            },
-        )
-
-        items = api_response.body.get('items', [])  # pyright: ignore[reportUnknownVariableType]
-        if len(items) == 0:  # pyright: ignore[reportUnknownArgumentType]
-            raise FileNotFoundError(
-                f'File not found in ICA: {parent_folder_path}{file_name}',
-            )
-        if len(items) > 1:  # pyright: ignore[reportUnknownArgumentType]
-            logger.warning(
-                f"Found multiple files named '{file_name}'; using the first one.",
-            )
-
-        file_id = items[0]['data'].get('id')  # pyright: ignore[reportUnknownVariableType]
-        if not file_id:
-            raise ValueError(f"Found file item for '{file_name}' but it has no ID.")
-
-        logger.info(f'Found file ID: {file_id}')
-        return file_id  # pyright: ignore[reportUnknownVariableType]
-
-    except icasdk.ApiException as e:
-        logger.error(f"API Error finding file '{file_name}': {e}")
-        raise
-    except Exception as e:
-        logger.error(f"Error finding file '{file_name}': {e}")
-        raise
-
-
 def get_md5_from_ica(
-    api_instance: project_data_api.ProjectDataApi,
+    api_instance: 'project_data_api.ProjectDataApi',
     path_parameters: dict[str, str],
     md5_file_id: str,
 ) -> tuple[str, str]:
@@ -369,11 +127,11 @@ def get_md5_from_ica(
 
 
 def stream_ica_file_to_gcs(
-    api_instance: project_data_api.ProjectDataApi,
+    api_instance: 'project_data_api.ProjectDataApi',
     path_parameters: dict[str, str],
     file_id: str,
     file_name: str,
-    gcs_bucket: Bucket,
+    gcs_bucket: 'Bucket',
     gcs_prefix: str,
     expected_md5_hash: str | None = None,
 ) -> None:
@@ -448,7 +206,7 @@ def stream_ica_file_to_gcs(
 
 
 def list_and_filter_ica_files(
-    api_instance: project_data_api.ProjectDataApi,
+    api_instance: 'project_data_api.ProjectDataApi',
     path_parameters: dict[str, str],
     base_ica_folder_path: str,
 ) -> list[tuple[str, str]]:
@@ -509,93 +267,8 @@ def list_and_filter_ica_files(
     return files_to_download
 
 
-def stream_files_to_gcs(
-    api_instance: project_data_api.ProjectDataApi,
-    path_parameters: dict[str, str],
-    files_to_download: list[tuple[str, str]],
-    gcs_bucket: Bucket,
-    gcs_output_path_prefix: str,
-) -> None:
-    """
-    Streams files from ICA pre-signed URLs directly to GCS blobs.
-    (Used by download_ica_pipeline_outputs.py)
-    """
-    bucket_name = gcs_bucket.name  # pyright: ignore[reportUnknownVariableType]
-    for file_name, file_id in files_to_download:
-        # Define the full GCS path for the file
-        gcs_blob_path = f'{gcs_output_path_prefix}/{file_name}'
-        blob = gcs_bucket.blob(gcs_blob_path)
-
-        logger.info(
-            f'Streaming {file_name} (ID: {file_id}) to gs://{bucket_name}/{gcs_blob_path}',
-        )
-
-        try:
-            # Get a pre-signed URL
-            url_response = api_instance.create_download_url_for_data(  # pyright: ignore[reportUnknownVariableType]
-                path_params=path_parameters | {'dataId': file_id},  # type: ignore[reportArgumentType]
-            )
-            download_url = url_response.body['url']  # pyright: ignore[reportUnknownVariableType]
-
-            # Download and upload as a stream
-            with requests.get(
-                download_url,
-                stream=True,
-                timeout=300,
-            ) as r:  # pyright: ignore[reportUnknownVariableType, reportUnknownArgumentType]
-                r.raise_for_status()
-                # r.raw is the raw byte stream (a file-like object)
-                blob.upload_from_file(
-                    r.raw,
-                    timeout=300,
-                )  # pyright: ignore[reportUnknownArgumentType]
-
-        except icasdk.ApiException as e:
-            logger.error(
-                f'Failed to get download URL for {file_name} (ID: {file_id}): {e}',
-            )
-        except requests.RequestException as e:
-            logger.error(
-                f'Failed to stream/download {file_name} (ID: {file_id}): {e}',
-            )
-        except gcs_exceptions.GoogleCloudError as e:
-            logger.error(f'An error occurred uploading to GCS for {file_name}: {e}')
-
-
-def get_file_details_from_ica(
-    api_instance: project_data_api.ProjectDataApi,
-    path_params: dict[str, str],
-    ica_folder_path: str,
-    file_name: str,
-) -> dict[str, Any] | None:
-    """
-    Checks if a file exists in ICA and returns its 'data' block if found.
-    (Used by upload_data_to_ica.py)
-    """
-    try:
-        query_params: dict[str, Any] = {
-            'parentFolderPath': ica_folder_path,
-            'filename': [file_name],
-            'filenameMatchMode': 'EXACT',
-            'pageSize': '2',
-        }
-
-        api_response = api_instance.get_project_data_list(
-            path_params=path_params,
-            query_params=query_params,
-        )
-        items = api_response.body.get('items', [])
-        if len(items) > 0:
-            return items[0]['data']  # pyright: ignore[reportUnknownVariableType]
-
-    except icasdk.ApiException as e:
-        logger.error(f'API error checking for file {file_name}: {e}')
-        # Don't raise, just return None
-    return None
-
-
 def check_file_existence(
-    api_instance: project_data_api.ProjectDataApi,
+    api_instance: 'project_data_api.ProjectDataApi',
     path_params: dict[str, str],
     ica_folder_path: str,
     cram_name: str,
@@ -605,7 +278,7 @@ def check_file_existence(
     (Used by upload_data_to_ica.py)
     """
     logger.info(f'Checking existence of {cram_name}...')
-    cram_data = get_file_details_from_ica(
+    cram_data = ica_api_utils.get_file_details_from_ica(
         api_instance,
         path_params,
         ica_folder_path,
@@ -616,60 +289,8 @@ def check_file_existence(
     return None
 
 
-def perform_upload_if_needed(cram_status: str | None, paths: dict[str, str]) -> None:
-    """
-    Handles the actual download from GCS and upload to ICA using CLIs.
-    (Used by upload_data_to_ica.py)
-    """
-    if cram_status == 'AVAILABLE':
-        logger.info(f'{paths["cram_name"]} already AVAILABLE in ICA. Skipping.')
-        return
-
-    # Authenticate ICA CLI
-    logger.info('Authenticating ICA CLI...')
-    # This command uses shell=True, but ICA_CLI_SETUP is a trusted constant
-    utils.run_subprocess_with_log(ICA_CLI_SETUP, 'Authenticate ICA CLI', shell=True)
-
-    local_dir = os.path.dirname(paths['local_cram_path'])
-    if not os.path.exists(local_dir):
-        os.makedirs(local_dir, exist_ok=True)
-        logger.info(f'Created local directory: {local_dir}')
-
-    # Download from GCS to local disk
-    logger.info(
-        f'Downloading {paths["cram_name"]} from GCS to {paths["local_cram_path"]}...',
-    )
-    utils.run_subprocess_with_log(
-        ['gcloud', 'storage', 'cp', paths['gcs_cram_path'], paths['local_cram_path']],
-        f'Download {paths["cram_name"]}',
-    )
-
-    logger.info(
-        f'Uploading {paths["local_cram_path"]} to ICA (using CLI for large file)...',
-    )
-    utils.run_subprocess_with_log(
-        [
-            'icav2',
-            'projectdata',
-            'upload',
-            paths['local_cram_path'],
-            paths['ica_folder_path'],
-        ],
-        f'Upload {paths["cram_name"]}',
-    )
-
-    # Clean up the large local file
-    try:
-        os.remove(paths['local_cram_path'])
-        logger.info(f'Removed local file: {paths["local_cram_path"]}')
-    except OSError as e:
-        logger.warning(
-            f'Could not remove local file {paths["local_cram_path"]}: {e}',
-        )
-
-
 def finalize_upload(
-    api_instance: project_data_api.ProjectDataApi,
+    api_instance: 'project_data_api.ProjectDataApi',
     path_params: dict[str, str],
     paths: dict[str, str],
     output_path_str: str,
@@ -679,7 +300,7 @@ def finalize_upload(
     (Used by upload_data_to_ica.py)
     """
     logger.info(f'Re-fetching file ID for {paths["sg_name"]}...')
-    cram_data = get_file_details_from_ica(
+    cram_data = ica_api_utils.get_file_details_from_ica(
         api_instance,
         path_params,
         paths['ica_folder_path'],
