+++ src/dragen_align_pa/jobs/download_md5_results.py
@@ -12,7 +12,7 @@ from cpg_utils.config import config_retrieve
 from icasdk.apis.tags import project_data_api
 from loguru import logger
 
-from dragen_align_pa import ica_utils
+from dragen_align_pa import ica_api_utils
 from dragen_align_pa.constants import BUCKET_NAME
 
 
@@ -24,14 +24,14 @@ def run(
     """
     Main function for the job.
     """
-    secrets: dict[Literal['projectID', 'apiKey'], str] = ica_utils.get_ica_secrets()
+    secrets: dict[Literal['projectID', 'apiKey'], str] = ica_api_utils.get_ica_secrets()
     project_id: str = secrets['projectID']
 
     path_parameters: dict[str, str] = {'projectId': project_id}
 
     folder_path: str = f'/{BUCKET_NAME}/{config_retrieve(["ica", "data_prep", "output_folder"])}'
 
-    with ica_utils.get_ica_api_client() as api_client:
+    with ica_api_utils.get_ica_api_client() as api_client:
         api_instance = project_data_api.ProjectDataApi(api_client)
 
         # Get the ID
