+++ src/dragen_align_pa/jobs/monitor_dragen_pipeline.py
@@ -4,7 +4,7 @@ from cpg_utils.config import config_retrieve
 from icasdk.apis.tags import project_analysis_api
 from loguru import logger
 
-from dragen_align_pa import ica_utils
+from dragen_align_pa import ica_api_utils
 
 
 def run(ica_pipeline_id: str | dict[str, str], is_mlr: bool = False) -> str:
@@ -21,7 +21,7 @@ def run(ica_pipeline_id: str | dict[str, str], is_mlr: bool = False) -> str:
     Returns:
         str: The pipeline status
     """
-    secrets: dict[Literal['projectID', 'apiKey'], str] = ica_utils.get_ica_secrets()
+    secrets: dict[Literal['projectID', 'apiKey'], str] = ica_api_utils.get_ica_secrets()
     project_id: str = secrets['projectID']
 
     pipeline_id: str = ica_pipeline_id['pipeline_id'] if isinstance(ica_pipeline_id, dict) else ica_pipeline_id
@@ -29,7 +29,7 @@ def run(ica_pipeline_id: str | dict[str, str], is_mlr: bool = False) -> str:
     logger.info(
         f'Monitoring pipeline run {pipeline_id}',
     )
-    with ica_utils.get_ica_api_client() as api_client:
+    with ica_api_utils.get_ica_api_client() as api_client:
         api_instance = project_analysis_api.ProjectAnalysisApi(api_client)
         if not is_mlr:
             path_params: dict[str, str] = {'projectId': project_id}
@@ -40,7 +40,7 @@ def run(ica_pipeline_id: str | dict[str, str], is_mlr: bool = False) -> str:
                 ),
             }
 
-        return ica_utils.check_ica_pipeline_status(
+        return ica_api_utils.check_ica_pipeline_status(
             api_instance=api_instance,
             path_params=path_params | {'analysisId': pipeline_id},
         )
