+++ src/dragen_align_pa/jobs/prepare_ica_for_analysis.py
@@ -7,7 +7,7 @@ from cpg_utils.config import config_retrieve
 from icasdk.apis.tags import project_data_api
 from loguru import logger
 
-from dragen_align_pa import ica_utils
+from dragen_align_pa import ica_api_utils, ica_utils
 from dragen_align_pa.constants import BUCKET_NAME
 
 
@@ -26,7 +26,7 @@ def run(
     Returns:
         dict [str, str] noting the analysis ID.
     """
-    secrets: dict[Literal['projectID', 'apiKey'], str] = ica_utils.get_ica_secrets()
+    secrets: dict[Literal['projectID', 'apiKey'], str] = ica_api_utils.get_ica_secrets()
     project_id: str = secrets['projectID']
 
     path_parameters: dict[str, str] = {'projectId': project_id}
@@ -35,7 +35,7 @@ def run(
         ['ica', 'data_prep', 'output_folder'],
     )
 
-    with ica_utils.get_ica_api_client() as api_client:
+    with ica_api_utils.get_ica_api_client() as api_client:
         api_instance = project_data_api.ProjectDataApi(api_client)
         folder_path: str = f'/{BUCKET_NAME}/{ica_analysis_output_folder}'
         for sg_name in cohort.get_sequencing_group_ids():
